First stab at refactoring params
The aim is to replace parameters per customer/project/repo by a more generic system
where parameters are grouped and stored on customer level.  Parameters have label
filters which are then applied to the labels of each repo.  This commit it an attempt
to rework this without breaking the existing functionality (as an exercise for the
future).