Register/unregister for all events
The zmq event broker does not "count" the number of times a client registers
for a single filter.  So when registering multiple times for the same filter
(e.g. when watching for a job to end), and then unregistering once, the client
does not receive any events anymore.  Either we need to solve this on broker
side, by adding reference counts, or soms sort of unique id (probably the most
robust solution), or fix it on client side.  For now, we register for all events
and only unregister when no more listeners remain.