Disabled retry when saving cache for now
For some as yet unknown reason, invoking the retry fn in the save-cache! function
throws a ClassNotFound error, for clojure.lang.PersistentArrayMap.  When removing
the retry, it still throws an exception, but now because files are supposedly
open, which should not happen because the process that opens them should at that
moment be terminated.